pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "apscheduler>=3.11.0",
    "email-validator>=2.2.0",
    "flask-login>=0.6.3",
    "flask>=3.1.1",
    "flask-sqlalchemy>=3.1.1",
    "flask-limiter>=3.10.0",
    "gunicorn>=23.0.0",
    "jinja2>=3.1.6",
    "msal>=1.33.0",
    "psycopg2-binary>=2.9.10",
    "requests>=2.32.4",
    "sqlalchemy>=2.0.42",
    "werkzeug>=3.1.3",
    "openai>=1.98.0",
    "flask-wtf>=1.2.2",
    "itsdangerous>=2.2.0",
    "twilio>=9.8.1",
    "pytest>=8.0.0",
    "pytest-cov>=6.0.0",
    "pytest-flask>=1.3.0",
    "factory-boy>=3.3.0",
    "beautifulsoup4>=4.14.2",
    "woocommerce>=3.0.0",
    "google-analytics-data>=0.19.0",
]
